chore: refactor app structure
