+++ src/components/Event/EventDetails/EventDetails.tsx
@@ -2,7 +2,7 @@ import { CustomDialog } from '../../../shared/components/CustomDialog';
 import type { CustomDialogProps } from '../../../shared/components/CustomDialog';
 import { EventDetailsContent } from './EventDetailsContent';
 import { EventEditButton } from '../EventEditButton';
-import type { UserEvent } from '../../../interfaces/UserEvent';
+import type { UserEvent } from '../../../shared/interfaces/UserEvent';
 import type { FunctionComponent, ReactElement } from 'react';
 
 interface EventDetailsProps extends Omit<CustomDialogProps, 'content' | 'title'> {
