+++ src/shared/interfaces/UserEvent.ts
