+++ src/shared/utils/Nil.ts
