+++ src/shared/utils/isNotNil.ts
